requirements.txt
<<<<<<< HEAD
aiohttp==3.7.4.post0
altgraph==0.17
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==3.0.1
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==4.1.0
boto3==1.20.26
botocore==1.23.26
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
colorama==0.4.4
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
discord==1.7.3
discord.py==1.7.3
elo==0.1.1
entrypoints==0.4
executing==0.8.3
future==0.18.2
google==3.0.0
google-api-core==1.26.0
google-api-python-client==2.0.2
google-auth==1.27.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.3
googleapis-common-protos==1.53.0
httplib2==0.19.0
idna==2.10
importlib-resources==5.4.0
ipykernel==6.9.1
ipython==8.1.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
mistune==0.8.4
mpmath==1.2.1
multidict==5.2.0
mysql-connector-python==8.0.23
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.22.2
oauthlib==3.1.0
packaging==20.9
pandocfilters==1.5.0
parso==0.8.3
pefile==2019.4.18
pickleshare==0.7.5
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.15.4
psycopg2-binary==2.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
pynder==0.0.13
pyparsing==2.4.7
pyrsistent==0.18.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
pywin32==303
pywin32-ctypes==0.2.0
pywinpty==2.0.2
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
requests==2.25.1
requests-oauthlib==1.3.0
roman==3.3
rsa==4.7.2
s3transfer==0.5.0
Send2Trash==1.8.0
six==1.15.0
soupsieve==2.2
stack-data==0.2.0
sympy==1.9
terminado==0.13.1
testpath==0.6.0
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.1
uritemplate==3.0.1
urllib3==1.26.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
yarl==1.7.2
zipp==3.7.0
=======
aiohttp==3.7.4.post0
altgraph==0.17
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==3.0.1
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==4.1.0
boto3==1.20.26
botocore==1.23.26
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
colorama==0.4.4
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
discord==1.7.3
discord.py==1.7.3
elo==0.1.1
entrypoints==0.4
executing==0.8.3
future==0.18.2
google==3.0.0
google-api-core==1.26.0
google-api-python-client==2.0.2
google-auth==1.27.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.3
googleapis-common-protos==1.53.0
httplib2==0.19.0
idna==2.10
importlib-resources==5.4.0
ipykernel==6.9.1
ipython==8.1.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
mistune==0.8.4
mpmath==1.2.1
multidict==5.2.0
mysql-connector-python==8.0.23
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.22.2
oauthlib==3.1.0
packaging==20.9
pandocfilters==1.5.0
parso==0.8.3
pefile==2019.4.18
pickleshare==0.7.5
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.15.4
psycopg2-binary==2.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
pynder==0.0.13
pyparsing==2.4.7
pyrsistent==0.18.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
pywin32==303
pywin32-ctypes==0.2.0
pywinpty==2.0.2
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
requests==2.25.1
requests-oauthlib==1.3.0
roman==3.3
rsa==4.7.2
s3transfer==0.5.0
Send2Trash==1.8.0
six==1.15.0
soupsieve==2.2
stack-data==0.2.0
sympy==1.9
terminado==0.13.1
testpath==0.6.0
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.1
uritemplate==3.0.1
urllib3==1.26.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
yarl==1.7.2
zipp==3.7.0
>>>>>>> 553bbca9c93d2d459bea8626b0be1181da904b4e
